Add support for custom timestamp name and field to nginx parser
Additionally, this refactors some commonly used time components into a
new package, httime. All parsers can now count on a universal 'Nower'
interface instead of re-declaring their own and timestamp parsing logic
(with respect to getting the fields from events) is contained in one
common location.
